Extract CUM_EXP_TABLE
